Make Item prefab and rename PlayerInventory class to Player
