+++ Assets/Scripts/Player/Player.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5dbd06f3dcba60a448974ba005f6ed3e
+guid: 97c54b74f999549e6beddc207e5aff55
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
